syscfg: Remove personal comments and such
Since the code is not personal anymore, we should, you know, cut all
opinions and such. Professionalism or something. Yeah, we just pretend
to not debug with printf statements like 'fk', 'wtf', ..., we do not
write "funny" stuff for code that shouldn't activate, et cetera. We
totally don't do that (AND THIS COMMIT MESSAGE IS TOTALLY PROFESSIONAL,
by the way).

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>