_autostart(): Use simple and static unit files
This patch left me wondering which is more garbage: the unit files of
Arch Linux for various programs or the init system `systemd`. Well, I
would say both are deserving of the title.

The reason is that converting the unit files to a simple "oneshot"
approach solved numerous problems caused by the complexity of `systemd`,
and everything just works now. I am not even taking into account the
slight but noticeable decrease in boot time.

Now, I do understand that there are two perspectives of the approach.
The simplicity might have reduced some functionality of `dnsmasq`, and
the security might not be as tight. However, it could also be the reason
for the elimination of certain problems that existed previously.

I would rather have a simple and stupid configuration than something as
complex as Arch Linux's systemd unit files. See, the former is at least
debuggable.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>