_sway_conf(): [systemd] write error handler
So, this garbage excuse of an init system cannot do its core purpose,
which is to start daemons and applications on boot. Or maybe it can,
but Arch Linux developers cannot write proper unit files.

See, wherever the issue resides, the core issue is clear: systemd is a
complex piece of... software. Let's pretend issues do not exit as long
as problematic services are restarted.

I try, day after day, to make progress towards a systemd-free system.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>