libfile(): Rewrite for `bash` support
BASH, as much as some people want it to be, is not fully
POSIX-compatible. That aside, the parser that BASH utilizes chokes on
nested case statements inside command substitution (variable command
substitution, at least).

Here, the syntax above is avoided by declaring the code inside functions
on the global stack. The downside obviously being that the sub functions
of libfile() are now global and needlessly polute the stack of syscfg.

To summarize: PLEASE, STOP LINKING /bin/sh TO /bin/bash! I SHOULD NOT
HAVE TO INTENTIONALLY REWRITE WITH LESS PERFORMANCE JUST TO SUPPORT A
SHELL THAT SHALL NOT BE CONSIDERED POSIX IN THE FIRST PLACE.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>