syscfg: Remove main_chk_home()
There is nothing inherently wrong with the function, but it is of
mediocre quality nonetheless.

But, a deeper reason for the removal is that $home (a syscfg equivalent
of $HOME) shall be an optional variable and location, and its clients
shall assert it themselves.

Of course, in the current iteration of syscfg, none of its
clients/consumers assert it (not directly, that is), but a precedent
shall be set.

Related: d6a58502f88c [1].

[1] syscfg: Remove main_chk_vars()

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>