_autostart(): [systemd] add portal impls to system.slice
On Arch Linux specifically at least, there is a weird disparity on what
software from XDG gets assigned to `system.slice` or is left assigned to
the default `app.slice`. There are both portal and portal
implementations both assigned and not assigned to `system.slice`,
respectively.

In our custom XDG units, we will assign all the portals and their
implementations to `system.slice` for simplicity. In practice, nothing
should be affected.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>