syscfg: Make $silent=0 default behavior
Only $silent support has been removed, while $silent_cmd still exists.
The reasons for the removal are quite apparent at just a quick glance.

The option in its current state features a lot of code duplication. But,
even if that were to be fully improved, the option is practically
useless, and its description rightfully says it's 'not recommended' to
enable it.

With the separation done in 6cf8fd960931 [1], coupled with the minimal
(and very useful) logging of syscfg, is $silent=1 support useless.

[1] syscfg: Separate syscfg and command logging

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>