_pipewire_conf(): Force 44.100 kHz
I care for audio, not for audio coming from videos (which usually is
48000k), and such audio is natively 44.100 kHz ("CD"), most of the
times. To avoid resampling in my context and preferred audio, force
44.100 kHz, to guide PipeWire to only resample the audio of general
media not meant to be listened to.

The aforementioned is supposedly as explained, but I am pretty stupid in
this regard, so I might be wrong. I tried my best to understand it all
and I would gladly accept informed corrections and suggestions.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>