syscfg: Remove $backup support
First, the implementation of this option is AWFUL. Second, the support
of this option is inherently flawed and we should instead focus on
streamlining the process of accepting/denying writes by showing a diff
and prompting the action to take based on it. Currently, none of the
targeted alternatives are supported, but a precedent is set.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>