syscfg: `chown 0700` our temp dir
`0700` is read, write, and execute only for root, not its group and
others (`0777`).

Like the files are set equivalently with `0600` octal mode, set our
temporary directory with `0700` octal mode. Here, this practice does not
result in any significant changes, but is a good practice nonetheless,
as it allows us to not depend on the default mode mask.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>