syscfg: Launch clients in a subshell
This is only a preliminary step that sets a precedent for clients.

At some point in the future, the clients of syscfg will be all external,
have their own PID (as they will be a different process), and be lexed
and parsed by syscfg.

Ideally, the shell code in syscfg clients will be restricted to what
syscfg allows and/or is configured to allow. This roughly implies that
clients will only be allowed to execute the library that syscfg is
configured to expose and nothing more.

The aforementioned aims to make syscfg a robust, secure POSIX shell
framework for system administration and configuration; one that is
easily auditable and provides a powerful toolkit to make system
configuration (and reconfiguration) a breeze.

Important considerations for syscfg clients that currently are not being
asserted, but might be in the future (this is not an extensive list of
requirements and is not yet official):
* Clients are encouraged to only use the syscfg library to achieve their
  goal.
* Clients are encouraged to put file creation and preparation logic
  (i.e. FILE_PATH() and FILE() calls, but not __file_write() calls)
  in own functions for each file. If the file itself has any additional
  conditional logic to it (e.g. removal on external conditions),
  the logic shall be declared in a different function.
* Client own functions shall be prefixed with `___` or risk being
  rejected, depending on how syscfg is configured.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>